+++ src/main/java/com/project/expenseTracker/repository/UserProfileRepository.java
@@ -3,5 +3,5 @@ package com.project.expenseTracker.repository;
 import com.project.expenseTracker.model.UserProfileInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserProfileRepository extends JpaRepository<UserProfileInfo, Integer> {
+public interface UserProfileRepository extends JpaRepository<UserProfileInfo, Long> {
 }
